+++ packages/google_vision/lib/src/authorization_exception.dart
@@ -1,9 +1,14 @@
 class AuthorizationException implements Exception {
   final String msg;
-  final Exception? err;
+  final Exception? _innerError;
 
-  const AuthorizationException([this.msg = '', this.err]);
+  const AuthorizationException([this.msg = '', Exception? innerError])
+    : _innerError = innerError;
 
   @override
-  String toString() => 'Authorization error: $msg\n\n$err';
+  String toString() => 'Authorization error: $msg';
+
+  /// Internal detailed error message for logging purposes only.
+  /// Do not expose to end users.
+  String toDetailedString() => 'Authorization error: $msg\n\n$_innerError';
 }