+++ packages/google_vision/lib/src/google_vision_file.dart
@@ -31,6 +31,10 @@ class GoogleVisionFile {
     List<int>? pages,
     String? parent,
   }) async {
+    RequestValidator.validateInputConfig(inputConfig);
+    RequestValidator.validateAnnotationType(annotationType);
+    RequestValidator.validateMaxResults(maxResults, 'maxResults');
+
     final batchAnnotateFilesResponse = await annotate(
       requests: [
         AnnotateFileRequest(